加入libuv，替换一部分socket相关处理。

+++ third/include/libuv/uv.h
@@ -292,7 +292,7 @@ typedef void (*uv_alloc_cb)(uv_handle_t* handle,
                             size_t suggested_size,
                             uv_buf_t* buf);
 typedef void (*uv_read_cb)(uv_stream_t* stream,
-                           ssize_t nread,
+	long nread,
                            const uv_buf_t* buf);
 typedef void (*uv_write_cb)(uv_write_t* req, int status);
 typedef void (*uv_connect_cb)(uv_connect_t* req, int status);
@@ -574,7 +574,7 @@ enum uv_udp_flags {
 
 typedef void (*uv_udp_send_cb)(uv_udp_send_t* req, int status);
 typedef void (*uv_udp_recv_cb)(uv_udp_t* handle,
-                               ssize_t nread,
+                               long nread,
                                const uv_buf_t* buf,
                                const struct sockaddr* addr,
                                unsigned flags);
@@ -1111,7 +1111,7 @@ struct uv_fs_s {
   uv_fs_type fs_type;
   uv_loop_t* loop;
   uv_fs_cb cb;
-  ssize_t result;
+  long result;
   void* ptr;
   const char* path;
   uv_stat_t statbuf;  /* Stores the result of uv_fs_stat() and uv_fs_fstat(). */
